Choropleth

Add points to a globe.

globe_choropleth(globe, data, country, cap_color = NULL,
  side_color = NULL, altitude = NULL, label = NULL,
  transition = 1000L, on_click = NULL, on_right_click = NULL,
  on_hover = NULL, match = c("auto", "iso2", "iso3", "name"))

Arguments

globe
An object of class globe as returned by create_globe, or an object of class globeProxy as returned by globeProxy.

data
A data.frame of points to draw.

country
The Bare column name containing either the ISO2, ISO3, or country name (see match argument).

cap_color
Bare column name containing the color of the surface.

side_color
Bare column name containing color of the sides.

altitude
Altitude of countries where (0 = 0 altitude (flat polygon), 1 = globe radius).

label
Bare column name containing label.

transition
A value of 0 will size the cone immediatel to their final altitude. New polygons are animated by rising them from the ground up.

on_click, on_right_click, on_hover
JavaScript functions as strings.

match
The type of country identifier, auto attempts to infer the type, iso2 (e.g.: "US"), iso3 (e.g.: "USA"), or the country name.

Examples

create_globe() %>%
  globe_img_url() %>%
  globe_choropleth(agriland, country_code, cap_color = percent)