Points

Add points to a globe.

globe_labels(globe, data, lat = NULL, lon = NULL, color = NULL,
  text = NULL, label = NULL, altitude = NULL, size = NULL,
  rotation = NULL, include_dot = NULL, dot_radius = NULL,
  dot_orientation = NULL, resolution = 2L, type_face = NULL,
  transition = 1000L, on_click = NULL, on_right_click = NULL,
  on_hover = NULL)

Arguments

globe
An object of class globe as returned by create_globe, or an object of class globeProxy as returned by globeProxy.

data
A data.frame of points to draw.

lat
Bare column names of points coordinates.

lon
Bare column names of points coordinates.

color
Bare column name of points color.

text
Bare column name containing text to display on globe.

label
Bare column name of points label. Supports plain text or HTML content.

altitude
Bare column name of points defining the cylinder's altitude in terms of globe radius units (0 = 0 altitude (flat circle), 1 = globe radius).

size
Bare column name of or the label text height, in angular degrees.

rotation
Bare column name for the label rotation in degrees. The rotation is performed clockwise along the axis of its latitude parallel plane.

include_dot
Bare column name for for whether to include a dot marker next to the text indicating the exact lat, lon coordinates of the label. If enabled the text will be rendered offset from the dot.

dot_radius
Bare column name for the radius of the dot marker, in angular degrees.

dot_orientation
Bare column name for the orientation of the label if the dot marker is present. Possible values are right, top and bottom.

resolution
Getter/setter for the radial geometric resolution of each cylinder, expressed in how many slice segments to divide the circumference. Higher values yield smoother cylinders.

type_face
Typeface to use for labels. Supports any typeface font generated by Facetype.js.

transition
Duration (ms) of the transition to animate point changes involving geometry modifications. A value of 0 will move the objects immediately to their final position. New objects are animated by scaling them from the ground up. Only works if merge is disabled.

on_click
JavaScript functions as strings.

on_right_click
JavaScript functions as strings.

on_hover
JavaScript functions as strings.

Examples

create_globe() %>%
  globe_img_url() %>%
  globe_labels(
    quakes,
    lat,
    long,
    text = stations
  )

library(shiny)

ui <- fluidPage(
  actionButton("add", "add quakes labels"),
  globeOutput("globe")
)

server <- function(input, output){
  output$globe <- renderGlobe({
    create_globe() %>%
      globe_img_url()
  })

  observeEvent(input$add, {
    globeProxy("globe") %>%
      globe_labels(
        quakes,
        lat,
        long,
        text = stations
      )
  })
}
# NOT RUN {
shinyApp(ui, server)
# }